Add support for adding untranslated keys
We have a requirement to be able to create
error messages which are just meant to be logged on server
and not translated.

However, the current model.AppError logic mandates an error id
which always must have a translation attached to it.

We circumvent this by adding a dummy id which is ignored by the
tool when looking for strings. Any code needing to create an
error without a translation can just use the dummy id.

While here, we remove pointers to a map :)